move module logic to modules folder
