+++ src/app/views/settings/SovereignClouds.tsx
@@ -8,13 +8,13 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   getCloudProperties, getCurrentCloud, globalCloud,
   replaceBaseUrl, storeCloudValue
-} from '../../../../modules/cloud-resolver';
-import { IRootState } from '../../../../types/root';
-import { setActiveCloud } from '../../../services/actions/cloud-action-creator';
-import { setSampleQuery } from '../../../services/actions/query-input-action-creators';
-import { setQueryResponseStatus } from '../../../services/actions/query-status-action-creator';
-import { translateMessage } from '../../../utils/translate-messages';
-import { Sovereign } from './cloud-options';
+} from '../../../modules/sovereign-clouds';
+import { IRootState } from '../../../types/root';
+import { setActiveCloud } from '../../services/actions/cloud-action-creator';
+import { setSampleQuery } from '../../services/actions/query-input-action-creators';
+import { setQueryResponseStatus } from '../../services/actions/query-status-action-creator';
+import { translateMessage } from '../../utils/translate-messages';
+import { Sovereign } from '../../../modules/sovereign-clouds/cloud-options';
 
 
 export const SovereignClouds = ({ cloudSelectorOpen, toggleCloudSelector, prompt = false }: any) => {
